+++ .isort.cfg
@@ -1,2 +1,2 @@
 [settings]
-known_third_party = PyInquirer,boto3,botocore,click,dateutil,halo,pytest,requests,setuptools
+known_third_party = PyInquirer,boto3,botocore,click,halo,moto,pytest,requests,setuptools
