+++ README.rst
@@ -6,8 +6,10 @@ Secure EC2
 .. image:: https://img.shields.io/pypi/v/secure_ec2.svg
         :target: https://pypi.python.org/pypi/secure_ec2
 
-.. image:: https://github.com/avishayil/secure_ec2/actions/workflows/test/badge.svg
-        :target: https://github.com/avishayil/secure_ec2/actions/workflows/test
+.. image:: https://github.com/avishayil/secure_ec2/actions/workflows/test.yml/badge.svg
+        :target: https://github.com/avishayil/secure_ec2/actions/workflows/test.yml
+
+.. image:: coverage.svg
 
 .. image:: https://readthedocs.org/projects/secure-ec2/badge/?version=latest
         :target: https://secure-ec2.readthedocs.io/en/latest/?version=latest
