+++ secure_ec2/secure_ec2.py
@@ -1,5 +1,5 @@
 """Main module."""
-from secure_ec2.src.aws import (
+from secure_ec2.src.api import (
     create_security_group,
     get_default_vpc,
     get_latest_ami,
@@ -8,11 +8,20 @@ from secure_ec2.src.aws import (
 )
 
 
-def main_(os_type: str, num_instances: int, keypair: str, instance_type: str):
-    image_id = get_latest_ami(os_type=os_type)
-    vpc_id = get_default_vpc()
-    subnet_id = get_subnet_id()
-    security_group_id = create_security_group(vpc_id=vpc_id, os_type=os_type)
+def main_(
+    os_type: str,
+    num_instances: int,
+    keypair: str,
+    instance_type: str,
+    profile: str,
+    region: str,
+) -> None:
+    image_id = get_latest_ami(os_type=os_type, profile=profile, region=region)
+    vpc_id = get_default_vpc(profile=profile, region=region)
+    subnet_id = get_subnet_id(profile=profile, region=region)
+    security_group_id = create_security_group(
+        vpc_id=vpc_id, os_type=os_type, profile=profile, region=region
+    )
     provision_ec2_instance(
         image_id=image_id,
         subnet_id=subnet_id,
@@ -20,4 +29,6 @@ def main_(os_type: str, num_instances: int, keypair: str, instance_type: str):
         keypair=keypair,
         instance_type=instance_type,
         num_instances=num_instances,
+        profile=profile,
+        region=region,
     )
