+++ tests/conftest.py
@@ -1,30 +1,39 @@
-from unittest.mock import patch
-
 import pytest
-from botocore.stub import Stubber
+from moto import mock_ec2, mock_iam
 
 from secure_ec2.src import helpers
-from tests.mocks import mock_get_ip_address
+
+
+def mock_get_ip_address():
+    return "192.168.1.1"
+
 
 pytest.MonkeyPatch().setattr(helpers, "get_ip_address", mock_get_ip_address)
-from secure_ec2.src.aws import ec2_client, iam_client  # noqa: E402
+from secure_ec2.src.aws import get_boto3_client, get_boto3_resource  # noqa: E402
 
 
 @pytest.fixture(autouse=True)
 def ec2_client_stub():
-    with Stubber(ec2_client) as stubber:
-        yield stubber
-        stubber.assert_no_pending_responses()
+    with mock_ec2():
+        ec2_client = get_boto3_client(
+            service="ec2", profile="default", region="us-east-1"
+        )
+        yield ec2_client
 
 
 @pytest.fixture(autouse=True)
 def ec2_resource_stub():
-    with patch("secure_ec2.src.aws.ec2_resource", return_value={}):
-        yield
+    with mock_ec2():
+        ec2_resource = get_boto3_resource(
+            service="ec2", profile="default", region="us-east-1"
+        )
+        yield ec2_resource
 
 
 @pytest.fixture(autouse=True)
 def iam_client_stub():
-    with Stubber(iam_client) as stubber:
-        yield stubber
-        stubber.assert_no_pending_responses()
+    with mock_iam():
+        iam_client = get_boto3_client(
+            service="iam", profile="default", region="us-east-1"
+        )
+        yield iam_client
