+++ tests/test_api.py
@@ -1,0 +1,66 @@
+from secure_ec2.src.api import (
+    create_security_group,
+    create_ssm_instance_profile,
+    get_default_vpc,
+    get_key_pairs,
+    get_latest_ami,
+    get_subnet_id,
+    provision_ec2_instance,
+)
+from secure_ec2.src.constants import SSM_ROLE_NAME
+
+
+def test_get_key_pairs(ec2_client_stub):
+    key_pairs = get_key_pairs(profile="default", region="us-east-1")
+    assert isinstance(key_pairs, list)
+
+
+def test_get_subnet_id(ec2_client_stub):
+    subnet_id = get_subnet_id(profile="default", region="us-east-1")
+    assert isinstance(subnet_id, str)
+
+
+def test_get_default_vpc(ec2_client_stub):
+    vpc_id = get_default_vpc(profile="default", region="us-east-1")
+    assert isinstance(vpc_id, str)
+
+
+def test_get_latest_ami(ec2_client_stub):
+    image_id = get_latest_ami(os_type="Demo", profile="default", region="us-east-1")
+    assert isinstance(image_id, str)
+
+
+def test_create_security_group(ec2_client_stub):
+    security_group_id = create_security_group(
+        vpc_id="vpc-123abcd", os_type="Windows", profile="default", region="us-east-1"
+    )
+    assert isinstance(security_group_id, str)
+
+
+def test_provision_ec2_instance(ec2_client_stub, ec2_resource_stub):
+    ec2_client_stub.create_key_pair(KeyName="demo-kp")
+    image_id = get_latest_ami(os_type="Demo", profile="default", region="us-east-1")
+    subnet_id = get_subnet_id(profile="default", region="us-east-1")
+    security_group_id = create_security_group(
+        vpc_id="vpc-123abcd", os_type="Windows", profile="default", region="us-east-1"
+    )
+    key_pair = get_key_pairs(profile="default", region="us-east-1")[0]
+    instance_id = provision_ec2_instance(
+        image_id=image_id,
+        subnet_id=subnet_id,
+        security_group_id=security_group_id,
+        keypair=key_pair,
+        instance_type="t2.micro",
+        num_instances=1,
+        profile="default",
+        region="us-east-1",
+    )
+    assert isinstance(instance_id, str)
+
+
+def test_create_ssm_instance_profile(iam_client_stub):
+    instance_profile = create_ssm_instance_profile(
+        profile="default", region="us-east-1"
+    )
+    assert isinstance(instance_profile, str)
+    assert instance_profile == SSM_ROLE_NAME
