+++ tests/test_helpers.py
@@ -1,0 +1,32 @@
+import re
+
+from secure_ec2.src.helpers import (
+    get_connection_port,
+    get_ip_address,
+    get_os_regex,
+    get_username,
+)
+
+
+def test_get_connection_port():
+    windows_connection_port = get_connection_port(os_type="Windows")
+    linux_connection_port = get_connection_port(os_type="Linux")
+    assert windows_connection_port == 3389
+    assert linux_connection_port == 22
+
+
+def test_get_username():
+    username = get_username()
+    assert isinstance(username, str)
+
+
+def test_get_ip_address():
+    ip_address = get_ip_address()
+    assert re.match(r"^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$", ip_address)
+
+
+def test_get_os_regex():
+    windows_os_regex = get_os_regex(os_type="Windows")
+    linux_os_regex = get_os_regex(os_type="Linux")
+    assert windows_os_regex == "Windows_Server-2019-English-Full-Base-*"
+    assert linux_os_regex == "amzn2-ami-hvm-2.0*"
