+++ tox.ini
@@ -1,21 +1,18 @@
 [tox]
 envlist = py38, flake8
 
-[travis]
-python =
-    3.8: py38
-
 [testenv:flake8]
 basepython = python
 deps = flake8
 commands = flake8 secure_ec2 tests
 
-[testenv]
+[testenv:py38]
 setenv =
     PYTHONPATH = {toxinidir}
     AWS_DEFAULT_REGION= us-east-1
 deps =
-    -r{toxinidir}/requirements_dev.txt
+    -e.[develop]
 commands =
     pip install -U pip
-    pytest --basetemp={envtmpdir} -vv
+    pytest --basetemp={envtmpdir} -vv --cov-report term --cov-report html:htmlcov --cov-report xml --cov={toxinidir}
+    coverage-badge -o coverage.svg -f
